LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/anonymous-client.https-expected.txt

CONSOLE MESSAGE: Prefetch request denied: not same origin as document
CONSOLE MESSAGE: Prefetch request denied: not same origin as document

FAIL test anonymous-client url prefetch for cross origin pages assert_equals: expected (string) "prefetch;anonymous-client-ip" but got (undefined) undefined
